Make a new component named "General"
- "GeneralContainer"는 일반적인 설정을 할 수 있다.
- 일반설정이란, 배너의 가로/세로 비율, 커스텀 크기 등이 있다.
