+++ src/SettingContainer.jsx
@@ -1,11 +1,15 @@
 import React from 'react';
 
+import GeneralContainer from './GeneralContainer';
 import BackgroundContainer from './BackgroundContainer';
 import ContentContainer from './ContentContainer';
 
 const SettingContainer = () => (
   <section>
+    <GeneralContainer />
+    <hr />
     <BackgroundContainer />
+    <hr />
     <ContentContainer />
   </section>
 );
